+++ extension-style/src/main/java/com/mapbox/maps/extension/style/light/Light.kt
@@ -2,7 +2,6 @@ package com.mapbox.maps.extension.style.light
 
 import com.mapbox.bindgen.Value
 import com.mapbox.common.toValue
-import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.MapboxStyleException
 import com.mapbox.maps.Style
 import com.mapbox.maps.extension.style.layers.properties.PropertyValue
@@ -12,12 +11,10 @@ import com.mapbox.maps.logE
 import kotlin.collections.HashMap
 
 /**
- * Global Light for 3D rendering. Must contain exactly one directional light and one ambient light.
- *
+ * Base class for all light implementations.
  * Check the [online documentation](https://www.mapbox.com/mapbox-gl-style-spec/#light).
  */
-@MapboxExperimental
-abstract class Lights3D {
+abstract class Light internal constructor() {
   /**
    * The ID of the Light.
    */
@@ -53,7 +50,7 @@ abstract class Lights3D {
       propertyValue.value
     )
     expected?.error?.let {
-      throw MapboxStyleException("Set lights3D property failed: $it")
+      throw MapboxStyleException("Set light property failed: $it")
     }
   }
 
@@ -62,12 +59,12 @@ abstract class Lights3D {
       return try {
         it.getStyleLightProperty(lightId, propertyName).unwrap()
       } catch (e: RuntimeException) {
-        logE(TAG, "Get lights3D property $propertyName failed: ${e.message}")
+        logE(TAG, "Get light property $propertyName failed: ${e.message}")
         logE(TAG, it.getStyleLightProperty(lightId, propertyName).toString())
         null
       }
     }
-    throw MapboxStyleException("Get property $propertyName failed: lights3D is not added to style yet.")
+    throw MapboxStyleException("Get property $propertyName failed: light is not added to style yet.")
   }
 
   internal fun setProperty(property: PropertyValue<*>) {
@@ -79,22 +76,22 @@ abstract class Lights3D {
     delegate?.let {
       return try {
         @Suppress("UNCHECKED_CAST")
-        val styleLights3DProperty =
+        val styleLightProperty =
           it.getStyleLightProperty(lightId, transitionName).value.contents as HashMap<String, Value>
-        val duration = styleLights3DProperty["duration"]?.contents as Long
-        val delay = styleLights3DProperty["delay"]?.contents as Long
+        val duration = styleLightProperty["duration"]?.contents as Long
+        val delay = styleLightProperty["delay"]?.contents as Long
         StyleTransition.Builder().delay(delay).duration(duration).build()
       } catch (e: RuntimeException) {
-        logE(TAG, "Get lights3D property failed: ${e.message}")
+        logE(TAG, "Get light property failed: ${e.message}")
         logE(TAG, it.getStyleLightProperty(lightId, transitionName).toString())
         null
       }
     }
-    throw MapboxStyleException("Get property $transitionName failed: lights3D is not added to style yet.")
+    throw MapboxStyleException("Get property $transitionName failed: light is not added to style yet.")
   }
 
   /**
-   * Get a string representation of the 3D Lights.
+   * Get a string representation of the Light.
    *
    * @return the string representation
    */
@@ -106,6 +103,6 @@ abstract class Lights3D {
    * Static variables and methods.
    */
   companion object {
-    private const val TAG = "Mbgl-Lights3D"
+    private const val TAG = "Mbgl-Light"
   }
 }