+++ extension-style/src/main/java/com/mapbox/maps/extension/style/light/LightExt.kt
@@ -1,3 +1,5 @@
+@file:JvmName("LightUtils")
+
 package com.mapbox.maps.extension.style.light
 
 import com.mapbox.bindgen.Value
@@ -7,67 +9,82 @@ import com.mapbox.maps.MapboxStyleException
 import com.mapbox.maps.Style
 import com.mapbox.maps.extension.style.layers.properties.PropertyValue
 import com.mapbox.maps.extension.style.light.generated.AmbientLight
+import com.mapbox.maps.extension.style.light.generated.AmbientLightDslReceiver
 import com.mapbox.maps.extension.style.light.generated.DirectionalLight
+import com.mapbox.maps.extension.style.light.generated.DirectionalLightDslReceiver
+import com.mapbox.maps.extension.style.light.generated.FlatLight
 import com.mapbox.maps.extension.style.utils.silentUnwrap
 import com.mapbox.maps.logE
 import kotlin.collections.HashMap
 
 /**
- * Extension function to add 3D Lights to the Style.
+ * Extension function to add dynamic light to the [Style].
+ * dynamic light is built from [AmbientLight] and [DirectionalLight].
  *
  * @param [ambientLight] The ambient light to be added
  * @param [directionalLight] The directional light to be added
- *
  */
 @MapboxExperimental
-fun Style.setup3DLights(ambientLight: AmbientLight, directionalLight: DirectionalLight) {
+fun Style.setLight(ambientLight: AmbientLight, directionalLight: DirectionalLight) {
   val ambientLightParam = HashMap<String, Value>().apply {
     this[LIGHT_PROPERTIES] = convertPropertyMapToValue(ambientLight.lightProperties)
     this.putAll(ambientLight.internalLightProperties)
   }
-
   val directionalLightParam = HashMap<String, Value>().apply {
     this[LIGHT_PROPERTIES] = convertPropertyMapToValue(directionalLight.lightProperties)
     this.putAll(directionalLight.internalLightProperties)
   }
-
   ambientLight.delegate = this
   directionalLight.delegate = this
-
   val valueList = ArrayList<Value>()
   valueList.add(Value(ambientLightParam))
   valueList.add(Value(directionalLightParam))
   val expected = this.setStyleLights(Value(valueList))
   expected.error?.let {
-    throw MapboxStyleException("Set lights3D failed: $it")
+    throw MapboxStyleException("Set dynamic light failed with error: $it")
   }
 }
 
 /**
- * Extension function to add list of 3D lights.
- * Note: Only [AmbientLight] and [DirectionalLight] are supported.
- * Lights must contain one instance of [AmbientLight] and one instance [DirectionalLight].
+ * Extension function to add the [FlatLight] to the [Style].
  *
- * @param list of [Lights3D]
+ * @param [flatLight] The flat light to be added
+ */
+fun Style.setLight(flatLight: FlatLight) {
+  val flatLightParam = HashMap<String, Value>().apply {
+    this[LIGHT_PROPERTIES] = convertPropertyMapToValue(flatLight.lightProperties)
+    this.putAll(flatLight.internalLightProperties)
+  }
+  flatLight.delegate = this
+  val valueList = ArrayList<Value>()
+  valueList.add(Value(flatLightParam))
+  val expected = this.setStyleLights(Value(valueList))
+  expected.error?.let {
+    throw MapboxStyleException("Set flat light failed with error: $it")
+  }
+}
+
+/**
+ * Extension function to add list of lights.
+ *
+ * Note: not all variations of lights is supported.
+ *
+ * @param lights list of [Light]
  */
 @MapboxExperimental
-fun Style.addLights3D(lights: List<Lights3D>) {
+fun Style.setLights(lights: List<Light>) {
   val valueList = ArrayList<Value>()
   for (light in lights) {
-    if (isLightSupported(light)) {
-      val lightParams = HashMap<String, Value>().apply {
-        this[LIGHT_PROPERTIES] = convertPropertyMapToValue(light.lightProperties)
-        this.putAll(light.internalLightProperties)
-      }.toValue()
-      light.delegate = this
-      valueList.add(lightParams)
-    } else {
-      throw MapboxStyleException("Light type $light not supported.")
-    }
+    val lightParams = HashMap<String, Value>().apply {
+      this[LIGHT_PROPERTIES] = convertPropertyMapToValue(light.lightProperties)
+      this.putAll(light.internalLightProperties)
+    }.toValue()
+    light.delegate = this
+    valueList.add(lightParams)
   }
   val expected = this.setStyleLights(Value(valueList))
   expected.error?.let {
-    throw MapboxStyleException("Set lights3D failed: $it")
+    throw MapboxStyleException("setStyleLights failed with error: $it")
   }
 }
 
@@ -82,15 +99,16 @@ private fun convertPropertyMapToValue(property: HashMap<String, PropertyValue<*>
 /**
  * Extension function to get 3D Light that has been applied to the style.
  *
- * @param lightId Id of 3D light.
+ * @param lightId Id of dynamic light.
  */
 @MapboxExperimental
-fun Style.getLights3D(lightId: String): Lights3D? {
+fun Style.getLight(lightId: String): Light? {
   return when (val type = getStyleLightProperty(lightId, "type").silentUnwrap<String>()) {
     "ambient" -> AmbientLight(lightId)
     "directional" -> DirectionalLight(lightId)
+    "flat" -> FlatLight(lightId)
     else -> {
-      logE(TAG, "Lights3D type: $type unknown.")
+      logE(TAG, "Light type: $type unknown.")
       null
     }
   }?.also { result ->
@@ -98,9 +116,18 @@ fun Style.getLights3D(lightId: String): Lights3D? {
   }
 }
 
+/**
+ * DSL function for creating [DynamicLight] instance.
+ */
 @MapboxExperimental
-private fun isLightSupported(light: Lights3D) = light is AmbientLight || light is DirectionalLight
+fun dynamicLight(
+  blockAmbient: AmbientLightDslReceiver.() -> Unit,
+  blockDirectional: DirectionalLightDslReceiver.() -> Unit
+): DynamicLight = DynamicLight(
+  AmbientLight("ambient").apply(blockAmbient),
+  DirectionalLight("directional").apply(blockDirectional)
+)
 
 @PublishedApi
-internal const val TAG = "Mbgl-Lights3DUtils"
+internal const val TAG = "Mbgl-LightUtils"
 internal const val LIGHT_PROPERTIES = "properties"